moved adding event listeners to a separate file
