+++ generatePage.js
@@ -17,9 +17,8 @@
 import UserData from './UserData.js';
 import { DEFAULT_DATA } from './defaultData.js';
 import {us, china, germany} from './demo.js';
-// import {elementStr, blockStr, findIcon} from './stringify.js';
 import { createPage, updateLocation, updatePage } from './createPage.js';
-import { viewPopup, closePopup, viewNewsTab, viewFinanceTab } from './eventHandlers.js';
+import {addEverything} from './eventListeners.js';
 // import * from './eventHandlers';
 // let user;
 // let lastSession = localStorage.getItem('lastSession');
@@ -29,6 +28,7 @@ document.onreadystatechange = () => {
   let state = document.readyState;
   // let user = new UserData(DEFAULT_DATA);
   let user = new UserData(china);
+  let backup = user;
   console.log(user);
 
   if (state === 'loading') {
@@ -40,94 +40,15 @@ document.onreadystatechange = () => {
   }
   else if (state === 'complete') {
     let blankState = document.body.innerHTML;
-    createPage(user);
+    let mainContent = document.querySelector('main').innerHTML;
+    createPage(user); 
+    addEverything(user, mainContent);
+    //so for some reason user becomes undefined, so redefine it???
+    
+    // const mainContent = document.querySelector('main').innerHTML;
+    // createPage(user);
     // console.log(user);
 
-    /**************************************EVENT LISTENERS*******************************************/
-
-    // PRIMARY NAV------------------------------------------------------------------------------------
-    // slide primary nav out of view when user closes it
-    document.querySelector('.primary-close').addEventListener('click', () => {
-      let nav = document.querySelector('.primary-nav');
-      nav.style.animation = 'slideOutNav 1s';
-      setTimeout(() => nav.classList.add('inactive-nav'), 1000);
-    });
-
-    // add event listener for each link in primary nav to slide popup into view
-    document.querySelectorAll('.link').forEach(link => {
-      link.addEventListener('click', () => {
-        viewPopup(link.innerText.toLowerCase());
-        document.querySelector('.primary-nav').classList.add('inactive-nav');
-      });
-    });
-
-    //slide primary nav into view when user clicks menu image
-    document.querySelector('.menu').addEventListener('click', () => {
-      document.querySelector('.primary-nav').style.animation = 'slideInNav 1s';
-      document.querySelector('.inactive-nav').classList.remove('inactive-nav');
-    });
-
-    //refresh page
-    document.querySelector('.refresh').addEventListener('click', async () => {
-      await user.refreshPage();
-      updatePage(blankState, user);
-      // document.body.innerHTML = blankState;
-
-      console.log(user);
-      // document = scaffold;
-      // createPage(user);
-      // updatePage(user);
-      // console.log(document.body.innerHTML);
-      // console.log(blankState);
-    })
-
-
-    // WIDGETS------------------------------------------------------------------------------------
-    // show selected widget from main screen
-    document.querySelectorAll('.widgets > section').forEach(widget => {
-      widget.addEventListener('click', () => viewPopup(widget.className.split('-')[0] + '-popup'));
-    });
-
-
-    // POPUPS + ACTIONS--------------------------------------------------------------------------
-    // makes close button in each popup and action section work
-    document.querySelectorAll('.close').forEach(button => {
-      button.addEventListener('click', () => closePopup(document.querySelector('.active-popup')));
-    });
-
-    // show selected finance tab
-    document.querySelectorAll('.finance-nav > li').forEach(tab => {
-      let className = tab.classList[0];
-      tab.addEventListener('click', () => viewFinanceTab(className, className + '-info'));
-    });
-
-    // make crypto the default active finance tab
-    document.querySelector('.crypto').classList.add('active-finance-tab');
-    document.querySelector('.crypto-info').classList.add('active-finance');
-
-    // show news tab
-    document.querySelectorAll('.news-nav > li').forEach(tab => {
-      let className = tab.classList[0];
-      tab.addEventListener('click', () => viewNewsTab(className, className + '-headlines'));
-    });
-
-    // make world headlines the default active news tab
-    document.querySelector('.national-headlines').classList.add('active-headlines');
-    document.querySelector('.national').classList.add('active-news-tab');
-
-
-    document.querySelector('.location-btn').addEventListener('click', () => {
-      console.log(user.location.coordinates);
-      if ('geolocation' in navigator) {
-        navigator.geolocation.getCurrentPosition(async position => {
-          let [lat, lon] = [position.coords.latitude, position.coords.longitude];
-          await user.reverseGeocode(lat, lon).then(res => console.log(res));
-          document.querySelector('#map > div').remove();
-          updateLocation(user.location);
-          // console.log(user.location);
-        });
-      }
-    });
 
   }
 }
